+++ KernelBench/level1/66_conv_standard_3D__asymmetric_input__asymmetric_kernel.py
@@ -17,7 +17,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: tuple, stride: tuple = (1, 1, 1), padding: tuple = (0, 0, 0), dilation: tuple = (1, 1, 1), groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv3d = nn.Conv3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias)
+        self.conv3d = nn.Conv3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.bfloat16)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -41,7 +41,7 @@ height = 128
 width = 128
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, depth, height, width)
+    x = torch.rand(batch_size, in_channels, depth, height, width, dtype=torch.bfloat16)
     return [x]
 
 def get_init_inputs():
