+++ KernelBench/level1/7_Matmul_with_small_K_dimension_.py
@@ -26,8 +26,8 @@ N = 16384 * 2
 K = 32 * 2
 
 def get_inputs():
-    A = torch.rand(M, K)
-    B = torch.rand(K, N)
+    A = torch.rand(M, K, dtype=torch.bfloat16)
+    B = torch.rand(K, N, dtype=torch.bfloat16)
     return [A, B]
 
 def get_init_inputs():
