+++ KernelBench/level1/86_conv_depthwise_separable_2D.py
@@ -16,8 +16,8 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: int, stride: int = 1, padding: int = 0, dilation: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.depthwise = nn.Conv2d(in_channels, in_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=in_channels, bias=bias)
-        self.pointwise = nn.Conv2d(in_channels, out_channels, kernel_size=1, bias=bias)
+        self.depthwise = nn.Conv2d(in_channels, in_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=in_channels, bias=bias, dtype=torch.bfloat16)
+        self.pointwise = nn.Conv2d(in_channels, out_channels, kernel_size=1, bias=bias, dtype=torch.bfloat16)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -45,7 +45,7 @@ padding = 1
 dilation = 1
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height, width)
+    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.bfloat16)
     return [x]
 
 def get_init_inputs():
