+++ KernelBench/level1/98_KLDivLoss.py
@@ -20,7 +20,7 @@ dim = 1
 
 def get_inputs():
     scale = torch.rand(())
-    return [(torch.rand(batch_size, *input_shape)*scale).softmax(dim=-1), torch.rand(batch_size, *input_shape).softmax(dim=-1)]
+    return [(torch.rand(batch_size, *input_shape, dtype=torch.bfloat16)*scale).softmax(dim=-1), torch.rand(batch_size, *input_shape, dtype=torch.bfloat16).softmax(dim=-1)]
 
 def get_init_inputs():
     return []
